requirements.txt
# Backend API dependencies for BrightSUM
# Python 3.10+

# FastAPI framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database ORM
sqlmodel==0.0.22

# Settings management
pydantic-settings==2.6.1

# Machine Learning
scikit-learn==1.5.2
numpy==2.1.3
pandas
joblib

# Development
pytest==8.3.3
httpx==0.27.2

# DB encryption
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0

# Email validation
email-validator==2.1.0.post1

# Pin bcrypt for Windows/Passlib
bcrypt==4.1.2

# To get authorize to work (fastAPI) 
python-multipart==0.0.9